refactor(api): move verification code routes to standard rest names
